refactor(agent-status): introduce CodexToolStatusResolver (title family)
Extract Codex terminal-title / suppression reconciliation out of
WorklaneStore+Metadata into a stateless CodexToolStatusResolver that
mutates PaneAuxiliaryState and returns outcome flags (TitleReconcileOutcome,
ShellReturnOutcome) the store honors for ready-status side effects and
enrichment-task cancellation. Store helpers become thin delegators; the
updateMetadata call order and per-helper write-back points are unchanged.
Adds the PaneToolStatusResolving seam (Codex-only for now).
